refactor: organise methods in backend package
